BugFix related to number_of_derivatives()
Problem was that the number of derivatives was not known at the creation
time of the body expression of the expressiongraph function.
This lead to CachedType nodes assuming the wrong number of derivatives
later on and assuming that these derivatives towards a higher index
number where zero.

Solved by adding a resize_nr_of_derivatives() function everywere
and calling this function when creating the call object. (size is
only increased, never decreased, such that at the end, the body
expression is properly sized for all call objects used.