Bug fix related to chained expressiongraph functions.
A subtle and difficult to find bug occurred.

When one expressiongraph function calls another one, we have to
use the parameter-node of the outer function call before the evaluation
node is created in the inner function call.  Previously, during
construction of the evaluation node, the parameters were examined and
the data structures were established according to the dependencies of
these parameters.  However, because the evaluation node of the outer
loop does not exists when adding the argument, these dependencies
are dummy dependencies that do not correspond to the actual
dependencies.

Therefore, construction and initialization of the evaluation node is
post-poned to the first value() or derivative() call.  At that time,
all dependencies are known.

The unit test 'test_function14.lua' of expressiongraph_context_lua
guards against this bug.